requirements.txt
streamlit
groq
requests
python-dotenv
duckduckgo-search>=5.0
regex